BridJ: - fixed 32 bits @Ptr annotation for return values - fixed classpath issue for callbacks - added debug build script

+++ libraries/Runtime/BridJ/src/test/cpp/test/test.cpp
@@ -342,13 +342,13 @@ TEST_API const wchar_t* wstringCStr(std::wstring* s) {
 #include "OAIdl.h" // for VARIANT
 #include "Vfw.h" // for CAPDRIVERCAPS
 
-TEST_API long long __cdecl sizeOfVARIANT() {
+TEST_API size_t __cdecl sizeOfVARIANT() {
 	return sizeof(VARIANT);
 }
-TEST_API long long __cdecl sizeOfDECIMAL() {
+TEST_API size_t __cdecl sizeOfDECIMAL() {
 	return sizeof(DECIMAL);
 }
-TEST_API long long __cdecl sizeOfCAPDRIVERCAPS() {
+TEST_API size_t __cdecl sizeOfCAPDRIVERCAPS() {
 	return sizeof(CAPDRIVERCAPS);
 }
